3.3-permissions/api_with_restrictions/requirements.txt
django
djangorestframework
psycopg2-binary
django-filter
flake8